stundas

+++ core.js
@@ -3,7 +3,7 @@ const client = new Discord.Client();
 
 client.on('message', message => {
     if (message.author.id === '482586747201519617') {
-        if (message.content.startsWith('/stunda')) {
+        if (message.content.startsWith('/fizika')) {
             message.channel.send('@everyone');
             const embed = new Discord.MessageEmbed()
                 .setThumbnail('https://lh3.googleusercontent.com/proxy/bk3MS2-oJeoNu4GPT6-Cvn5GTt6q8MmO6gohbTYLBFVg1pl-wjBBTlX4NPjxJgDCe3Kpx6yoQesCSjF7wDkwovii2ichXU-BuAFeo1CIJ_d5')
@@ -14,6 +14,26 @@ client.on('message', message => {
                 .addFields(
                 { name: 'SĀKUMS: Tagad', value: 'Lai pievienotos stundai, spied uz zili iekrāsotā teksta', inline: true },
                 { name: 'Stundas tēma:', value: 'Praktiskais darbs - spēks un paātrinājums', inline: true }
+                );
+            message.channel.send(embed);
+            message.delete();
+        };
+    };
+});
+
+client.on('message', message => {
+    if (message.author.id === '482586747201519617') {
+        if (message.content.startsWith('/eikt')) {
+            message.channel.send('@everyone');
+            const embed = new Discord.MessageEmbed()
+                .setThumbnail('https://lh3.googleusercontent.com/proxy/bk3MS2-oJeoNu4GPT6-Cvn5GTt6q8MmO6gohbTYLBFVg1pl-wjBBTlX4NPjxJgDCe3Kpx6yoQesCSjF7wDkwovii2ichXU-BuAFeo1CIJ_d5')
+                .setTitle('EIKT nozares tehnisko darbu pamatiemaņas PA2')
+                .setAuthor('TIEŠSAISTES STUNDA')
+                .setURL('https://us05web.zoom.us/j/86872407614?pwd=L1FtejhWTG1IZjZPelV4ekxWYnd1QT09')
+                .setFooter('ID: 868 7240 7614, Passcode: zdXF4Z')
+                .addFields(
+                { name: 'SĀKUMS: 13:30', value: 'Lai pievienotos stundai, spied uz zili iekrāsotā teksta', inline: true },
+                { name: 'Stundas tēma:', value: 'N/A', inline: true }
                 );
             message.channel.send(embed);
             message.delete();
